fix(phase1): address second Codex review pass on PR #112
Six more P2 findings from the second automated review:

- LegacyRuntimeOffset: use LegacyValues.phpTrim for is_daylight_saving
  instead of String.trim(), so form-feed padding no longer flips the
  runtime UTC offset an hour off PHP's reading.
- LegacyScheduleCalendar.parseDaysOffToDows: same phpTrim fix for the
  schedule module's independent days_off parser.
- LegacyScheduleService/LegacyScheduleCalendar.monthOverview: keep
  month/year at 64-bit width instead of narrowing to int, so a value
  outside the 32-bit range no longer wraps into an unrelated valid
  month instead of clamping like PHP does.
- LegacyCheckInService.checkOut: return duration_minutes as a long
  instead of narrowing phpRound's result to int, so a sufficiently
  distant check-in no longer wraps into an unrelated positive number.
- LegacyRequestTypeService.create: require the post-insert re-read
  like the attendance create path already does, so a concurrent
  delete becomes the generic 500 D-084 expects rather than a 201 with
  no data key.
- application.properties: resolve multipart requests lazily so an
  oversized or malformed upload still reaches the method/auth/
  company/role/availability guards before failing, instead of being
  rejected by DispatcherServlet before the handler runs.

+++ backend/src/main/java/com/workin/legacy/attendance/records/LegacyCheckInService.java
@@ -80,7 +80,7 @@ public class LegacyCheckInService {
 	}
 
 	/** {@code ok(CHECK_OUT_RECORDED, ['duration_minutes' => ..., 'time' => ...])}. */
-	public record CheckOutOutcome(int durationMinutes, String time) {
+	public record CheckOutOutcome(long durationMinutes, String time) {
 	}
 
 	/**
@@ -195,7 +195,7 @@ public class LegacyCheckInService {
 		long seconds = checkIn == null
 				? 0L
 				: java.time.Duration.between(checkIn, clock.now()).getSeconds();
-		return new CheckOutOutcome((int) phpRound(seconds / 60d), clock.now().format(SQL_DATE_TIME));
+		return new CheckOutOutcome(phpRound(seconds / 60d), clock.now().format(SQL_DATE_TIME));
 	}
 
 	// ------------------------------------------------------------------
